daemon/requirements.txt
mss==9.0.1
imagehash==4.3.1
httpx==0.27.0
Pillow==10.2.0
PyYAML==6.0.1